refactor: code structure and improvements
